requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.3.1
Flask-Cors==5.0.0
python-dotenv==1.0.1
pyodbc==4.0.32
Werkzeug==3.0.3
SQLAlchemy==2.0.35
